Add input_path option to FastFrames.py
This option allows the user to specify the path to a directory
containing filelist.txt and sum_of_weights.txt files. This is
useful, for example, if multiple people are working on the
same FastFrames configuration and don't want to repeatedly
change the hard-coded input path in the config file.

I just added an input to the BlockReaderGeneral class that
passes along the input_path. The BlockReaderGeneral will first
read the filelist.txt and sum_of_weights.txt files from the
config (or set them to None if not specified) --- if
input_path has been provided, then these values will be overwritten.