Fix false "buggy file" warning in merge_empty_grid_files.py
The warning was caused whenever there was an attempt
to merge a single file. A guard clause returned early
from the merge function, but did not return a boolean
corresponding to the status (which should be vacuously
True). Thus, the caller thought that the merge failed.

Fixed the problem by adding a True return to the
guard clause.